chore: try moving lib to app
