+++ app/lib/get-wm-endpoints.js
