+++ app/lib/ignored-endpoints.js
